+++ dynamic_forms/fields.py
@@ -25,12 +25,19 @@ class TextMultiSelectField(six.with_metaclass(models.SubfieldBase,
     def contribute_to_class(self, cls, name):
         super(TextMultiSelectField, self).contribute_to_class(cls, name)
         if self.choices:
-            def _func(self, fieldname=name, choicedict=dict(self.choices)):
+            def _func(self, fieldname=name):
                 return self.separate_values_by.join([
-                    choicedict.get(value, value) for value in
+                    self.choices.get(value, value) for value in
                     getattr(self, fieldname)
                 ])
             setattr(cls, 'get_%s_display' % self.name, _func)
+
+    def deconstruct(self):
+        name, path, args, kwargs = super(TextMultiSelectField, self).deconstruct()
+        kwargs['separate_values_by'] = self.separate_values_by
+        if kwargs.get('separate_values_by', None) == '\n':
+            del kwargs['separate_values_by']
+        return name, path, args, kwargs
 
     def formfield(self, **kwargs):
         # don't call super, as that overrides default widget if it has choices