+++ tests/settings.py
@@ -16,9 +16,17 @@ INSTALLED_APPS = (
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.admin',
-    'dynamic_forms',
+    'captcha',
     'tests',
 )
+if django.VERSION[:2] < (1, 7):
+    INSTALLED_APPS = INSTALLED_APPS + (
+        'dynamic_forms',
+    )
+else:
+    INSTALLED_APPS = INSTALLED_APPS + (
+        'dynamic_forms.apps.DynamicFormsConfig',
+    )
 
 SECRET_KEY = 'test-secret-key'
 
@@ -31,10 +39,6 @@ MIDDLEWARE_CLASSES = (
     'dynamic_forms.middlewares.FormModelMiddleware',
 )
 
-MIGRATION_MODULES = {
-    'dynamic_forms': 'dynamic_forms.dj_migrations',
-}
-
 ROOT_URLCONF = 'tests.urls'
 
 TEMPLATE_DIRS = (
@@ -43,3 +47,12 @@ TEMPLATE_DIRS = (
 
 if django.VERSION[:2] < (1, 6):
     TEST_RUNNER = 'discover_runner.DiscoverRunner'
+
+DYNAMIC_FORMS_FORM_TEMPLATES = (
+    ('dynamic_forms/form.html', 'Default form template'),
+    ('template1.html', 'Test tempate 1'),
+)
+DYNAMIC_FORMS_SUCCESS_TEMPLATES = (
+    ('dynamic_forms/form_success.html', 'Default success template'),
+    ('template2.html', 'Test template 2'),
+)