+++ tests/urls.py
@@ -5,4 +5,7 @@ admin.autodiscover()
 
 urlpatterns = patterns('',
     url(r'^admin/', include(admin.site.urls)),
+    url(r'^dynamic_forms/',
+        include('dynamic_forms.urls', namespace='dynamic_forms')),
+    url(r'^captcha/', include('captcha.urls')),
 )